requirements.txt
flask>=2.0.1
numpy>=1.21.0
opencv-python-headless>=4.5.3
pillow>=8.3.1
torch>=1.9.0
torchvision>=0.10.0
werkzeug>=2.0.1
gunicorn>=20.1.0
python-dotenv>=0.19.0
transformers>=4.30.0
flask-cors>=3.0.10
gdown>=4.7.1
